Finish native performance tests

+++ subprojects/performance/src/templates/cpp-source/lib.h
@@ -1,6 +1,6 @@
 #ifndef PROJECT_HEADER_${functionName}_H
 #define PROJECT_HEADER_${functionName}_H
 <% functionCount.times { %>
-int CPP_${functionName}_${it + 1} ();
+long CPP_${functionName}_${it + 1} ();
 <% } %>
 #endif // PROJECT_HEADER_${functionName}_H
